Category-related code into categories.js
Left the UI manager only in main.js, and moved category-specific management into a specific class of its own
